requirements.txt
pillow
numpy
gradio